requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
cryptography==46.0.3
fastapi
filelock==3.20.0
firebase-admin
fsspec==2025.9.0
google-api-python-client
google-auth-httplib2
google-cloud-vision
h11==0.16.0
httplib2==0.31.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
numpy
packaging
pandas
pdfminer.six==20250506
pdfplumber
pillow==12.0.0
psutil==7.1.0
puremagic==1.30
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pypandoc
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
tqdm
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn
psycopg2-binary==2.9.11
opencv-python-headless
python-docx==1.2.0
openpyxl==3.1.5
pikepdf
beautifulsoup4
google-generativeai>=0.5.0
google-cloud-documentai
pgvector